+++ packages/lit-element/src/lib/updating-element.ts
@@ -209,6 +209,9 @@ const defaultPropertyDeclaration: PropertyDeclaration = {
   hasChanged: notEqual,
 };
 
+export type connectCallback = () => void;
+export type updateCallback = (changedProperties: PropertyValues) => void;
+
 /**
  * The Closure JS Compiler doesn't currently have good support for static
  * property semantics where "this" is dynamic (e.g.
@@ -465,6 +468,26 @@ export abstract class UpdatingElement extends HTMLElement {
    */
   private _reflectingProperty: PropertyKey | null = null;
 
+  /**
+   * Set of callbacks called in connectedCallback.
+   */
+  connectedCallbacks: Set<connectCallback> = new Set();
+
+  /**
+   * Set of callbacks called in disconnectedCallback.
+   */
+  disconnectedCallbacks: Set<connectCallback> = new Set();
+
+  /**
+   * Set of callbacks called before update.
+   */
+  updateCallbacks: Set<updateCallback> = new Set();
+
+  /**
+   * Set of callbacks called after updated.
+   */
+  updatedCallbacks: Set<updateCallback> = new Set();
+
   constructor() {
     super();
     this.initialize();
@@ -512,6 +535,7 @@ export abstract class UpdatingElement extends HTMLElement {
     // Ensure first connection completes an update. Updates cannot complete
     // before connection.
     this.enableUpdating();
+    this.connectedCallbacks.forEach((cb) => cb());
   }
 
   /**
@@ -526,7 +550,9 @@ export abstract class UpdatingElement extends HTMLElement {
    * reserving the possibility of making non-breaking feature additions
    * when disconnecting at some point in the future.
    */
-  disconnectedCallback() {}
+  disconnectedCallback() {
+    this.disconnectedCallbacks.forEach((cb) => cb());
+  }
 
   /**
    * Synchronizes property values when attributes change.
@@ -708,6 +734,7 @@ export abstract class UpdatingElement extends HTMLElement {
     try {
       shouldUpdate = this.shouldUpdate(changedProperties);
       if (shouldUpdate) {
+        this.updateCallbacks.forEach((cb) => cb(changedProperties));
         this.update(changedProperties);
       } else {
         this._markUpdated();
@@ -727,6 +754,7 @@ export abstract class UpdatingElement extends HTMLElement {
         this.firstUpdated(changedProperties);
       }
       this.updated(changedProperties);
+      this.updatedCallbacks.forEach((cb) => cb(changedProperties));
     }
   }