+++ packages/lit-element/src/test/updating-element_test.ts
@@ -2623,4 +2623,178 @@ suite('UpdatingElement', () => {
     assert.equal(a.foo, 20);
     assert.equal(a.updatedFoo, 20);
   });
+
+  suite('lifecycle callbacks', () => {
+    class A extends UpdatingElement {
+      static properties = {foo: {}};
+      foo = 'foo';
+      updateCount = 0;
+      updatedCount = 0;
+      connectedCount = 0;
+      disconnectedCount = 0;
+
+      connectedCallback() {
+        this.connectedCount++;
+        super.connectedCallback();
+      }
+
+      disconnectedCallback() {
+        this.disconnectedCount++;
+        super.disconnectedCallback();
+      }
+
+      update(changedProperties: PropertyValues) {
+        this.updateCount++;
+        super.update(changedProperties);
+      }
+
+      updated(changedProperties: PropertyValues) {
+        this.updatedCount++;
+        super.updated(changedProperties);
+      }
+    }
+    customElements.define(generateElementName(), A);
+    let el!: A;
+
+    setup(async () => {
+      container = document.createElement('div');
+      document.body.appendChild(container);
+      el = new A();
+      container.appendChild(el);
+      await el.updateComplete;
+    });
+
+    teardown(() => {
+      if (container && container.parentNode) {
+        container.parentNode.removeChild(container);
+      }
+    });
+
+    test('connected/disconnected callbacks', () => {
+      let connectedCount = 0;
+      el.connectedCallbacks.add(() => connectedCount++);
+      let disconnectedCount = 0;
+      el.disconnectedCallbacks.add(() => disconnectedCount++);
+      assert.equal(el.connectedCount, 1);
+      assert.equal(el.disconnectedCount, 0);
+      assert.equal(connectedCount, 0);
+      assert.equal(disconnectedCount, 0);
+      container.removeChild(el);
+      assert.equal(el.connectedCount, 1);
+      assert.equal(el.disconnectedCount, 1);
+      assert.equal(connectedCount, 0);
+      assert.equal(disconnectedCount, 1);
+      container.appendChild(el);
+      assert.equal(el.connectedCount, 2);
+      assert.equal(el.disconnectedCount, 1);
+      assert.equal(connectedCount, 1);
+      assert.equal(disconnectedCount, 1);
+    });
+
+    test('update/updated callbacks', async () => {
+      let updateCount = 0;
+      let updateChangedPropeties: PropertyValues | null = null;
+      let updatedChangedPropeties: PropertyValues | null = null;
+      el.updateCallbacks.add((changedProperties: PropertyValues) => {
+        updateChangedPropeties = changedProperties;
+        updateCount++;
+      });
+      let updatedCount = 0;
+      el.updatedCallbacks.add((changedProperties: PropertyValues) => {
+        updatedChangedPropeties = changedProperties;
+        updatedCount++;
+      });
+      assert.equal(el.updateCount, 1);
+      assert.equal(el.updatedCount, 1);
+      assert.equal(updateCount, 0);
+      assert.equal(updatedCount, 0);
+      el.foo = 'new';
+      await el.updateComplete;
+      assert.equal(el.updateCount, 2);
+      assert.equal(el.updatedCount, 2);
+      assert.equal(updateCount, 1);
+      const expectedChangedProperties = new Map([['foo', 'foo']]);
+      assert.deepEqual(updateChangedPropeties, expectedChangedProperties);
+      assert.equal(updatedCount, 1);
+      assert.deepEqual(updatedChangedPropeties, expectedChangedProperties);
+    });
+
+    test('callbacks can only be added once', async () => {
+      let connectedCount = 0;
+      const connectedCallback = () => {
+        connectedCount++;
+      };
+      let disconnectedCount = 0;
+      const disconnectedCallback = () => {
+        disconnectedCount++;
+      };
+      let updateCount = 0;
+      const updateCallback = () => {
+        updateCount++;
+      };
+      let updatedCount = 0;
+      const updatedCallback = () => {
+        updatedCount++;
+      };
+      el.connectedCallbacks.add(connectedCallback);
+      el.connectedCallbacks.add(connectedCallback);
+      el.disconnectedCallbacks.add(disconnectedCallback);
+      el.disconnectedCallbacks.add(disconnectedCallback);
+      el.updateCallbacks.add(updateCallback);
+      el.updateCallbacks.add(updateCallback);
+      el.updatedCallbacks.add(updatedCallback);
+      el.updatedCallbacks.add(updatedCallback);
+      container.removeChild(el);
+      assert.equal(disconnectedCount, 1);
+      container.appendChild(el);
+      assert.equal(connectedCount, 1);
+      el.foo = 'foo2';
+      await el.updateComplete;
+      assert.equal(updateCount, 1);
+      assert.equal(updatedCount, 1);
+    });
+
+    test('callbacks can be removed', async () => {
+      let connectedCount = 0;
+      const connectedCallback = () => {
+        connectedCount++;
+      };
+      let disconnectedCount = 0;
+      const disconnectedCallback = () => {
+        disconnectedCount++;
+      };
+      let updateCount = 0;
+      const updateCallback = () => {
+        updateCount++;
+      };
+      let updatedCount = 0;
+      const updatedCallback = () => {
+        updatedCount++;
+      };
+      el.connectedCallbacks.add(connectedCallback);
+      el.disconnectedCallbacks.add(disconnectedCallback);
+      el.updateCallbacks.add(updateCallback);
+      el.updatedCallbacks.add(updatedCallback);
+      container.removeChild(el);
+      assert.equal(disconnectedCount, 1);
+      container.appendChild(el);
+      assert.equal(connectedCount, 1);
+      el.foo = 'foo2';
+      await el.updateComplete;
+      assert.equal(updateCount, 1);
+      assert.equal(updatedCount, 1);
+      el.connectedCallbacks.delete(connectedCallback);
+      el.disconnectedCallbacks.delete(disconnectedCallback);
+      el.updateCallbacks.delete(updateCallback);
+      el.updatedCallbacks.delete(updatedCallback);
+      container.removeChild(el);
+      container.appendChild(el);
+      el.foo = 'foo2';
+      await el.updateComplete;
+      assert.equal(disconnectedCount, 1);
+      assert.equal(connectedCount, 1);
+      assert.equal(updateCount, 1);
+      assert.equal(updatedCount, 1);
+    });
+  });
 });